More progress on fixing lists.
This removes the :li_span element type, and fixes how Maruku chooses to wrap list items in paragraphs in many cases (though not all). There are still quite a few list cases that still fail, but this fixes #26, fixes #30, fixes #64, fixes #67, and fixes #72, and activates many of the pending list tests.